feat(survey): surveyCreationのUIモジュール化リファクタリング
Issue #23 に基づき、アンケート作成画面 (surveyCreation.html) のリファクタリングを実施。

- ヘッダー、サイドバー、フッターを共通コンポーネントとして動的に読み込むように変更。
- アコーディオン、日付ピッカー、ドラッグ&ドロップのUIロジックをそれぞれ独立したモジュール (accordion.js, datepicker.js, draggable.js) に分割。

これにより、設計原則に準拠し、コードの再利用性と保守性が向上しました。
